% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(a, b = 0)
}
\arguments{
\item{a}{numeric; the first number}

\item{b}{numeric; the second number}
}
\value{
a single item numeric vector that is the sum of a and b
}
\description{
Add two numbers together with this function
}
\examples{

#add 3 and 4 together
add(3,4)

# one number adds 0
add(2)
}
