% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.r
\name{multiply}
\alias{multiply}
\title{Multiply two numbers}
\usage{
multiply(a, b = 1)
}
\arguments{
\item{a}{numeric; the first number}

\item{b}{numeric; the second number}
}
\value{
a single item numeric vector that is the multiplication of a and b
}
\description{
Multiply two numbers together with this function
}
\examples{

# multiply 3 and 4 together
multiply(3,4)

# one number multiplies by 1
multiply(2)
}
